created shared routes
